Rename 'dev' command to 'server'
- Rename 'dev' CLI command to 'server' for clarity
- Update function name from Dev to Server in server.go
- Remove old basic server command from commands_linux.go
- Update default cluster name from 'dev' to 'local'
- Update references in shell scripts and documentation
- Change default runner ID from 'dev' to 'runtime'
- Update log messages to reflect new terminology
- Fix errcheck linter errors

This provides a clearer, more operational naming convention for the runtime server command.
